+++ src/test/php/lang/PropertiesTest.php
@@ -722,7 +722,7 @@ class PropertiesTest extends \PHPUnit_Framework_TestCase
         $this->assertFalse($this->properties->parseBool('scalar', 'intValue2'));
         $this->assertFalse($this->properties->parseBool('scalar', 'floatValue1'));
         $this->assertFalse($this->properties->parseBool('scalar', 'floatValue2'));
-        $this->assertTrue($this->properties->parseBool('scalar', 'boolValue1'));
+        $this->assertFalse($this->properties->parseBool('scalar', 'boolValue1'));
         $this->assertTrue($this->properties->parseBool('scalar', 'boolValue2'));
         $this->assertTrue($this->properties->parseBool('scalar', 'boolValue3'));
         $this->assertTrue($this->properties->parseBool('scalar', 'boolValue4'));
@@ -767,7 +767,7 @@ class PropertiesTest extends \PHPUnit_Framework_TestCase
         $this->assertFalse($this->properties->parseBool('scalar', 'intValue2', true));
         $this->assertFalse($this->properties->parseBool('scalar', 'floatValue1', true));
         $this->assertFalse($this->properties->parseBool('scalar', 'floatValue2', true));
-        $this->assertTrue($this->properties->parseBool('scalar', 'boolValue1', true));
+        $this->assertFalse($this->properties->parseBool('scalar', 'boolValue1', true));
         $this->assertTrue($this->properties->parseBool('scalar', 'boolValue2', true));
         $this->assertTrue($this->properties->parseBool('scalar', 'boolValue3', true));
         $this->assertTrue($this->properties->parseBool('scalar', 'boolValue4', true));